modifiche in corso a lezioni ed esercizi insegnante

+++ insegnanti/elimina_lezione.php
@@ -6,10 +6,19 @@ $id = $_GET['id'];
 $result = $conn->query("SELECT * FROM lezione WHERE id='$id'");
 $lezione = $result->fetch_assoc();
 
-unlink("../" . $lezione['presentazione']);
-unlink("../" . $lezione['lezione']);
 
-$conn->query("DELETE FROM lezione  WHERE id='$id'");
+mysqli_autocommit($conn, FALSE);
+$conn->query("LOCK TABLES lezione WRITE");
+$conn->query("BEGIN");
+$r = $conn->query("DELETE FROM lezione  WHERE id='$id'");
+if ($r) {
+    unlink("../" . $lezione['presentazione']);
+    unlink("../" . $lezione['lezione']);
+    $conn->query("COMMIT"); 
+} else {
+    $conn->query("ROLLBACK");
+}
+$conn->query("UNLOCK TABLES");
 
 header("Location: ../nuova-lezione.html");
 ?>
